whisper/requirements.txt
openai-whisper
requests
minio